test_case_generator/pyproject.toml
[tool.poetry]
name = "test-case-generator"
version = "0.1.0"
description = "The Test Case Generator Agent is for analyzing business requirements documents and generating structured test cases."
authors = ["Mridula Vinod <mridulait67@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    {include = "testcase_agent"}
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.1"
google-adk = { version = "==1.8.0", extras = ["eval"] }
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.88.0"}
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { version = "==1.8.0", extras = ["eval"] }
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"
pandas = "^2.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
